fix: try to fix the BlockingIOError, see below message for details;
BlockingIOError: [Errno 35] Unable to create file (unable to lock file, errno = 35, error message = 'Resource temporarily unavailable')
This may be caused by the program creates h5 files for multiple times;
